chore(b6): split conformance_kat.rs into directory module helpers (closes #60)
595-LOC test file was past the project's 500-line guideline. Pure
structural refactor — no semantic change. Split into:

- conformance_kat.rs (entry): the two #[test] fns + the dispatch
  loop, ~194 LOC.
- conformance_kat_helpers/types.rs: KAT vector deserialization types
  + BridgeOrSyntheticErr wrapper.
- conformance_kat_helpers/fixtures.rs: kat_path, fixtures_dir, and
  the resolve_* family.
- conformance_kat_helpers/errors.rs: variant_name_vault,
  vault_error_detail, assert_err, read_block_err_* helpers.
- conformance_kat_helpers/dispatch.rs: per-op runners +
  assert_open_ok + assert_read_block_ok (kept inline rather than
  splitting further, since the bridge crate's per-record handle type
  is named Record not RecordHandle and inlining sidesteps the
  type-naming question).

Mirrors the existing core/tests/common/ pattern (mod.rs + sibling
.rs files). All existing test names, semantics, panic messages, and
counts are preserved unchanged. Gauntlet: 641 passed + 10 ignored
(unchanged); clippy clean; fmt OK.
